Added referrer for immediate assistance
The system now listens to the ?assistance referer to call the
go("assistance") function, which includes name generation and crowd control.
The "Immediate assistance" button on the homepage (https://tripsit.me/)
will link to this referrer and immediately call the assistance function,
helping people in need reach a free #tripsit channel faster.